refactor: move category-related components inside categories/
