+++ src/components/layout/categories/category-bar.tsx
